chore(packages/markdown): setup modularized approach for tailwind and components, integrate ellipsis from management frontend
