+++ packages/markdown/src/Ellipsis.tsx
@@ -1,12 +1,12 @@
-import Markdown from '@klicker-uzh/markdown'
 import { Prose, Tooltip } from '@uzh-bf/design-system'
 import React from 'react'
 import { twMerge } from 'tailwind-merge'
+import Markdown from './Markdown'
 
-interface EllipsisProps {
+export interface EllipsisBaseProps {
   children: string
   maxLength?: number
-  maxLines?: number
+  maxLines?: 1 | 2 | 3
   withoutPopup?: boolean
   className?: {
     root?: string
@@ -16,14 +16,16 @@ interface EllipsisProps {
   }
 }
 
-interface EllipsisPropsMaxLength extends EllipsisProps {
+export interface EllipsisPropsMaxLength extends EllipsisBaseProps {
   maxLength: number
   maxLines?: never
 }
-interface EllipsisPropsMaxLines extends EllipsisProps {
+export interface EllipsisPropsMaxLines extends EllipsisBaseProps {
   maxLength?: never
-  maxLines: number
+  maxLines: 1 | 2 | 3
 }
+
+export type EllipsisProps = EllipsisPropsMaxLength | EllipsisPropsMaxLines
 
 function Ellipsis({
   children,
@@ -31,17 +33,19 @@ function Ellipsis({
   maxLines,
   withoutPopup,
   className,
-}: EllipsisPropsMaxLength | EllipsisPropsMaxLines): React.ReactElement {
+}: EllipsisProps): React.ReactElement {
   if (maxLines) {
     return (
       <Tooltip
+        delay={3000}
         tooltip={
           <Prose
             className={{
-              root: 'flex-initial max-w-full leading-6 prose-p:m-0 prose-invert text-white hover:text-white prose-img:m-0',
+              root: 'flex-initial max-w-full leading-6 prose-p:m-0 hover:text-white prose-img:m-0 ',
             }}
           >
             <Markdown
+              withModal={false}
               content={children
                 .toString()
                 .replace(/^(- |[0-9]+\. |\* |\+ )/g, '')}
@@ -51,37 +55,32 @@ function Ellipsis({
         }
         className={{
           tooltip: twMerge(
-            'text-sm max-w-[50%] md:max-w-[60%]',
+            'text-sm bg-white border shadow max-w-md',
             className?.tooltip
           ),
         }}
         withIndicator={false}
       >
-        <div
-          className={twMerge(
-            'line-clamp-1 line-clamp-2 line-clamp-3',
-            // HACK: dynamic classnames do not work with tailwind
-            // the above ensures classes 1-3 are present
-            `line-clamp-${maxLines}`,
-            className?.root
-          )}
+        <Prose
+          className={{
+            root: twMerge(
+              'flex-initial max-w-full leading-6 prose-p:m-0 text-black hover:text-black prose-img:m-0',
+              // HACK: dynamic classnames do not work with tailwind
+              // line-clamp-1 line-clamp-2 line-clamp-3
+              // the above ensures classes 1-3 are present
+              `line-clamp-${maxLines}`,
+              className?.root,
+              className?.content
+            ),
+          }}
         >
-          <Prose
-            className={{
-              root: twMerge(
-                'flex-initial max-w-full leading-6 prose-p:m-0 text-black hover:text-black prose-img:m-0',
-                className?.content
-              ),
-            }}
-          >
-            <Markdown
-              content={children
-                .toString()
-                .replace(/^(- |[0-9]+\. |\* |\+ )/g, '')}
-              className={{ root: className?.markdown, img: 'max-h-36' }}
-            />
-          </Prose>
-        </div>
+          <Markdown
+            content={children
+              .toString()
+              .replace(/^(- |[0-9]+\. |\* |\+ )/g, '')}
+            className={{ root: className?.markdown, img: 'max-h-16' }}
+          />
+        </Prose>
       </Tooltip>
     )
   }
@@ -150,8 +149,10 @@ function Ellipsis({
         shortenedParsedContent
       ) : (
         <Tooltip
+          delay={3000}
           tooltip={
             <Markdown
+              withModal={false}
               content={children
                 .toString()
                 .replace(/^(- |[0-9]+\. |\* |\+ )/g, '')}
