+++ packages/markdown/src/components.css
@@ -1,2 +1,1 @@
 @tailwind components;
-@tailwind utilities;
