+++ packages/markdown/src/utilities.css
@@ -1,0 +1,1 @@
+@tailwind utilities;
